backend/requirements.txt
# Core
fastapi==0.115.5
uvicorn==0.34.0
python-dotenv==1.0.1
httpx==0.27.2

# Auth & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.1

# Database
psycopg2-binary==2.9.11

# AI / Workers
groq==0.15.0
google-generativeai==0.5.2
openai==1.12.0
ollama

# Files / Media
python-multipart==0.0.12
python-docx==1.1.2
PyPDF2==3.0.1
Pillow==10.4.0
reportlab==4.0.9
aiofiles==24.1.0

# System & Logging
psutil==5.9.8
python-json-logger==2.0.7
requests==2.32.3
